+++ source/dea-app/src/persistence/schema/dea-table.ts
@@ -6,11 +6,11 @@
 import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
 import { Table } from 'dynamodb-onetable';
 import { Dynamo } from 'dynamodb-onetable/Dynamo';
-import { getRequiredEnv } from '../../lambda-http-helpers';
+import { getCustomUserAgent, getRequiredEnv } from '../../lambda-http-helpers';
 import { DeaSchema } from './dea-schema';
 
 const region = process.env.AWS_REGION ?? 'us-east-1';
-const client = new Dynamo({ client: new DynamoDBClient({ region }) });
+const client = new Dynamo({ client: new DynamoDBClient({ region, customUserAgent: getCustomUserAgent() }) });
 const deaTableName = getRequiredEnv('TABLE_NAME', 'TABLE_NAME is not set in your lambda!');
 
 export const deaTable = new Table({