+++ source/dea-ui/ui/next.config.js
@@ -1,7 +1,7 @@
 const withTM = require('next-transpile-modules')(['@cloudscape-design/components']);
 
 const STAGE = process.env.NEXT_PUBLIC_STAGE ?? 'devsample';
-const USING_CUSTOM_DOMAIN = process.env.NEXT_PUBLIC_IS_USING_CUSTOM_DOMAIN ?? false;
+const USING_CUSTOM_DOMAIN = process.env.NEXT_PUBLIC_IS_USING_CUSTOM_DOMAIN?.trim().toLowerCase() === 'true';
 
 const basePath = USING_CUSTOM_DOMAIN ? `/ui` : `/${STAGE}/ui`;
 /** @type {import('next').NextConfig} */