+++ source/dea-ui/ui/src/helpers/apiHelper.ts
@@ -4,10 +4,10 @@
  */
 import { aws4Interceptor, Credentials } from 'aws4-axios';
 import axios, { AxiosError, AxiosRequestConfig } from 'axios';
+import { isUsingCustomDomain } from '../common/utility';
 import { refreshCredentials, signOutProcess } from './authService';
 
 let urlBase = process.env.NEXT_PUBLIC_DEA_API_URL;
-const isUsingCustomDomain = process.env.NEXT_PUBLIC_IS_USING_CUSTOM_DOMAIN ? true : false;
 if (typeof window !== 'undefined' && !urlBase) {
   if (isUsingCustomDomain) {
     urlBase = `https://${window.location.hostname}/`;