+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2023-08-15
+
+### Added
+- configuration to opt out of DynamoDB Dataplane events in CloudTrail
+
+### Fixed
+- bump semver and minimist packages
+- improve custom domain support
+- gov region one-click fix
+
 ## [1.0.3] - 2023-07-24
 
 ### Fixed
@@ -14,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - various UI/UX improvements
+- cdk deployment for us-gov-east-1
+- cloud formation template for us-gov-west-1
 
 ## [1.0.1] - 2023-06-30